Skip record marked as deleted
Each record is preceded by a deletion flag
that tells whether the record is considered deleted or not.

Previously we would read that flag but simply ignore it.

Now the reader should properly skip records marked as deleted.

Also the code was written in a way were we considered that the
flag could be an explict field of the record, but it is not possible and
we remove that part to simplify things.