Use bufrw for the BufReadWriteFile
The implementation of BufReadWriteFile used in order
to have buffered read and write on a std::fs::File
was made in a 'hacky' way.
It was using 2 files, one wrapped in a BufReader and the other
in a BufWriter, this impl was not completely correct.

Indeed there was a synchronization bug that was happening in release
build but not in debug build because a debug_assert called a method
that would force the synchronization.

So we replace this by BufReaderWriter type from the bufrw
crate which should not properly implement buffered read + write